+++ src/migration/ConfigMigrationHandlerInput.js
@@ -2,12 +2,13 @@ const ConfigMigrationHandler = require('./ConfigMigrationHandler');
 
 class ConfigMigrationHandlerInput extends ConfigMigrationHandler {
     get lastVersion() {
-        return 1; // Don't forget to update node declaration too
+        return 2; // Don't forget to update node declaration too
     }
 
     migrate(controller) {
         this.controller = controller;
         if (this.currentVersion === undefined) this.migrateFromLegacy();
+        if (this.currentVersion === 1) this.migrateHomeKitPayload();
         this.result.new.config_version = this.config_version;
     }
 
@@ -37,6 +38,17 @@ class ConfigMigrationHandlerInput extends ConfigMigrationHandler {
         this.result.delete.push('outputAtStartup');
 
         this.config_version = 1;
+    }
+
+    migrateHomeKitPayload() {
+        this.result.new.output_rules = this.config.output_rules.map((rule) => {
+            if (rule.type === 'homekit') {
+                this.result.info.push('node-red-contrib-deconz/server:tip.homekit_payload');
+                rule.payload = ['__auto__'];
+            }
+            return rule;
+        });
+        this.config_version = 2;
     }
 
 }